% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{bicluster_extension_process}
\alias{bicluster_extension_process}
\title{Bicluster extension process}
\usage{
bicluster_extension_process(MIR_profile, seed_row, seed_col, zero_ratio)
}
\arguments{
\item{MIR_profile}{A binarized microRNA profile}

\item{seed_row}{List of a seed row index}

\item{seed_col}{List of a seed column index}

\item{zero_ratio}{Zero ratio to be allowed.}
}
\value{
extended matrix
}
\description{
Bicluster extension process
}
