% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{bicluster_trimming_process}
\alias{bicluster_trimming_process}
\title{Bicluster trimming process}
\usage{
bicluster_trimming_process(table, zero_ratio)
}
\arguments{
\item{table}{A matrix that will be reduced}

\item{zero_ratio}{Zero ratio allowed.}
}
\value{
Reduced matrix
}
\description{
Bicluster trimming process
}
