% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{comatrix}
\alias{comatrix}
\title{Pariwise distance matrix for the entire list of biclusters}
\usage{
comatrix(biclusts, MIR_profile)
}
\arguments{
\item{biclusts}{Lists of biclusters.}

\item{MIR_profile}{binarized microRNA profile}
}
\value{
pariwise distance matrix
}
\description{
Pariwise distance matrix for the entire list of biclusters
}
