% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{expCondition}
\alias{expCondition}
\title{Get experimantal condition information}
\usage{
expCondition()
}
\value{
Experimental condition table
}
\description{
Get experimantal condition information
}
