\name{genelist}
\alias{genelist}
\docType{data}
\title{A list of genes}
\description{This is a list of 20639 gene symbols from GPL570.}
\usage{data("genelist")}
\format{
  The format is:
 chr "genelist"
}
\details{ [HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array }
\source{ [HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array }
\references{ [HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array }
\examples{
data(genelist)
## maybe str(genelist) ; plot(genelist) ...
}
\keyword{datasets}
