% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRprofile_from_FCtable.R
\name{get_miR_target}
\alias{get_miR_target}
\title{Get sequence-based miRNA targets}
\usage{
get_miR_target(miRNA)
}
\arguments{
\item{miRNA}{miRNA name. One of miRNAs listed in "miRNAlist" data.}
}
\value{
A vector of sequence-based miRNA targets
}
\description{
Returns miRNA targets predicted from three or more algorithms.
}
\examples{
get_miR_target('hsa-miR-1-3p')
}
