% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{get_zero_rate}
\alias{get_zero_rate}
\title{Get zero rate}
\usage{
get_zero_rate(target)
}
\arguments{
\item{target}{a matrix}
}
\value{
The ratio of zero in a binarized matrix
}
\description{
Get zero rate
}
