% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRprofile_from_FCtable.R
\name{getAllmiRNATargetIndex}
\alias{getAllmiRNATargetIndex}
\title{Load human miRNA target index}
\usage{
getAllmiRNATargetIndex()
}
\value{
A vector of human miRNA target index
}
\description{
human miRNA target index
}
