% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRprofile_from_FCtable.R
\name{getMIRprofile}
\alias{getMIRprofile}
\title{Get MIRprofile}
\usage{
getMIRprofile(miRNA, FCtable = NULL, FCcutoff = log2(1.3))
}
\arguments{
\item{miRNA}{miRNA name. One of miRNAs listed in "miRNAlist" data.}

\item{FCtable}{FC matrix of 20,639 genes for 5,158 conditions. If NULL, it will be loaded from the local.}

\item{FCcutoff}{Fold change cutoff in log2 scale. Default = log2(1.3). To extract MIR profile for target down-regulation in test condition, take negative value.}
}
\value{
A matrix of fold change value of sequence-based miRNA targets under selected cell conditions.
}
\description{
Returns MIR profile for input miRNA.
}
\examples{
fc=load_FCtable(); MIRprofile = getMIRprofile('hsa-miR-1-3p',fc,log2(1.3))
}
