% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRprofile_from_FCtable.R
\name{load_FCtable}
\alias{load_FCtable}
\title{load_FCtable}
\usage{
load_FCtable()
}
\value{
A fold change matrix of 5,158 rows and 20,639 columns.
}
\description{
Loading a large table of fold change values for 20,639 human genes in 5,158 cell conditions.
}
