% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{make_biclust}
\alias{make_biclust}
\title{Make all biclusters}
\usage{
make_biclust(MIR_PROFILE, FCcutoff, REPETITION, STEP_NUMBER, ZERORATE)
}
\arguments{
\item{MIR_PROFILE}{A matrix of MIR profile.}

\item{FCcutoff}{Binarization fold change cutoff in log2-scale for MIR_PROFILE. Target up-regulation biclusters will be created for positive FC cutoff, and down-regulation biclusters will be generated with negative FC cutoff.}

\item{REPETITION}{The number of running ensemble bicluster function.}

\item{STEP_NUMBER}{The number of extension process}

\item{ZERORATE}{Final zero rate allowed.}
}
\value{
list of seed and biclusters
}
\description{
Make all biclusters
}
