% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{merge_bicluster}
\alias{merge_bicluster}
\title{Merging similar biclusters}
\usage{
merge_bicluster(MIR_profile, tree_cutoff = 0.5, biclusts, seedlist,
  FCcutoff = log2(1.3), mir = "testmiRNA", ZERORATE)
}
\arguments{
\item{MIR_profile}{binarized microRNA profile}

\item{tree_cutoff}{Similarity cutoff}

\item{biclusts}{list of extended biclusters}

\item{seedlist}{list of seeds}

\item{FCcutoff}{Fold change cutoff (in log2 sclale) for binarization of MIR profile}

\item{mir}{Name of miRNA}

\item{ZERORATE}{Final zero ratio allowed}
}
\value{
Extended bicluster list
}
\description{
Merging similar biclusters
}
