\name{miRNAlist}
\alias{miRNAlist}
\docType{data}
\title{A list of microRNAs}
\description{These are a list of microRNA names.}
\usage{data("miRNAlist")}
\format{
  The format is:
 chr "miRNAlist"
}
\details{List of 2632 microRNAs.}
\source{This data was obtained from miRBase.}
\references{This data was obtained from miRBase.}
\examples{
data(miRNAlist)
## maybe str(miRNAlist) ; plot(miRNAlist) ...
}
\keyword{datasets}
