% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{PBE_MERGE}
\alias{PBE_MERGE}
\title{Progressive extension and merging of biclusters}
\usage{
PBE_MERGE(MIR_profile, mir, biclust.path = "./", FCcutoff = log2(1.3),
  REPETITION = 10, STEP_NUMBER = 10, ZERORATE = 0.1, tree_cutoff = 0.5)
}
\arguments{
\item{MIR_profile}{A matrix of microRNA profile where rows are cell conditions and columns are sequence-based miRNA targets.}

\item{mir}{Name of microRNA}

\item{biclust.path}{Directory where biclusts will be saved}

\item{FCcutoff}{Log2-fold change binarization cutoff for seed. Default = log2(1.3)}

\item{REPETITION}{The number of repetition that the ensemble function runs to extract seed biclusters. Default = 10}

\item{STEP_NUMBER}{The number of extension process. Default = 10}

\item{ZERORATE}{Final zero rate allowed. Default = 0.1}

\item{tree_cutoff}{Similarity cutoff. Default = 0.5}
}
\value{
Extended and merged biclusters in the directory assigned to ‘biclust.path’.
}
\description{
Progressive extension and merging of biclusters
}
