% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust_extension_merging.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity of two biclusters}
\usage{
similarity(biclust1, biclust2)
}
\arguments{
\item{biclust1}{First bicluster matrix}

\item{biclust2}{Second bicluster matrix}
}
\value{
Similarity of two biclusters
}
\description{
Calculate similarity of two biclusters
}
